[State Sync] Move SynchronizationState into coordinator.rs to better group logic.
