+++ state-sync/src/coordinator.rs
@@ -10,7 +10,6 @@ use crate::{
     logging::{LogEntry, LogEvent, LogSchema},
     network::{StateSynchronizerEvents, StateSynchronizerMsg, StateSynchronizerSender},
     request_manager::RequestManager,
-    state_synchronizer::SynchronizationState,
 };
 use anyhow::{bail, ensure, format_err, Result};
 use diem_config::{
@@ -20,10 +19,13 @@ use diem_config::{
 use diem_logger::prelude::*;
 use diem_mempool::{CommitResponse, CommittedTransaction};
 use diem_types::{
+    epoch_change::Verifier,
+    epoch_state::EpochState,
     ledger_info::LedgerInfoWithSignatures,
     transaction::{Transaction, TransactionListWithProof, Version},
     waypoint::Waypoint,
 };
+use executor_types::ExecutedTrees;
 use fail::fail_point;
 use futures::{
     channel::{mpsc, oneshot},
@@ -1340,6 +1342,72 @@ impl<T: ExecutorProxyTrait> SyncCoordinator<T> {
                 failed_msg
             )
         })
+    }
+}
+
+/// SynchronizationState contains the following fields:
+/// * `committed_ledger_info` holds the latest certified ledger info (committed to storage),
+///    i.e., the ledger info for the highest version for which storage has all ledger state.
+/// * `synced_trees` holds the latest transaction accumulator and state tree (which may
+///    or may not be committed to storage), i.e., some ledger state for the next highest
+///    ledger info version is missing.
+/// * `trusted_epoch_state` corresponds to the current epoch if the highest committed
+///    ledger info (`committed_ledger_info`) is in the middle of the epoch, otherwise, it
+///    corresponds to the next epoch if the highest committed ledger info ends the epoch.
+///
+/// Note: `committed_ledger_info` is used for helping other Diem nodes synchronize (i.e.,
+/// it corresponds to the highest version we have a proof for in storage). `synced_trees`
+/// is used locally for retrieving missing chunks for the local storage.
+#[derive(Clone)]
+pub struct SynchronizationState {
+    committed_ledger_info: LedgerInfoWithSignatures,
+    synced_trees: ExecutedTrees,
+    trusted_epoch_state: EpochState,
+}
+
+impl SynchronizationState {
+    pub fn new(
+        committed_ledger_info: LedgerInfoWithSignatures,
+        synced_trees: ExecutedTrees,
+        current_epoch_state: EpochState,
+    ) -> Self {
+        let trusted_epoch_state = committed_ledger_info
+            .ledger_info()
+            .next_epoch_state()
+            .cloned()
+            .unwrap_or(current_epoch_state);
+
+        SynchronizationState {
+            committed_ledger_info,
+            synced_trees,
+            trusted_epoch_state,
+        }
+    }
+
+    pub fn committed_epoch(&self) -> u64 {
+        self.committed_ledger_info.ledger_info().epoch()
+    }
+
+    pub fn committed_ledger_info(&self) -> LedgerInfoWithSignatures {
+        self.committed_ledger_info.clone()
+    }
+
+    pub fn committed_version(&self) -> u64 {
+        self.committed_ledger_info.ledger_info().version()
+    }
+
+    /// Returns the highest available version in the local storage, even if it's not
+    /// committed (i.e., covered by a ledger info).
+    pub fn synced_version(&self) -> u64 {
+        self.synced_trees.version().unwrap_or(0)
+    }
+
+    pub fn trusted_epoch(&self) -> u64 {
+        self.trusted_epoch_state.epoch
+    }
+
+    pub fn verify_ledger_info(&self, ledger_info: &LedgerInfoWithSignatures) -> Result<()> {
+        self.trusted_epoch_state.verify(ledger_info)
     }
 }
